+++ icedata/fridge/__init__.py
@@ -1,3 +1,4 @@
 from icedata.fridge.data import *
 from icedata.fridge.parsers import *
 from icedata.fridge.visualization import *
+from icedata.fridge import trained_models