+++ icedata/utils/__init__.py
@@ -1,0 +1,1 @@
+from icedata.utils.utils import *