+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 
 ## Added
+- Added `trained_models` module
 - `NUM_CLASSES` constant to datasets
 
 ## Changed